feat: touched the controllers, to give health its own as it will need to check database connectivity later
