+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "express": "4.17.1",
     "helmet": "3.21.2",
+    "mongoose": "^5.7.7",
     "server-side-tools": "git+https://github.com/wh-iterabb-it/server-side-tools.git"
   },
   "devDependencies": {
