+++ src/controllers/address.js
@@ -1,28 +1,6 @@
 const logger = require('server-side-tools').logger;
 
-class Controller {
-  constructor () {
-    // TODO:
-  }
-
-  messageResponse(msg) {
-    const response = {
-      'data': [],
-
-    };
-
-    response.data.push(msg);
-    return response;
-  }
-
-  /**
-   * @param {Request} req - Express request object
-   * @param {Response} res - Express response object
-   */
-  methodNotAllowed(req, res) {
-    res.status(405).json(this.messageResponse('Method Not Allowed'));
-  }
-}
+const { Controller } = require('./');
 
 class AddressController extends Controller {
   constructor () {
