+++ src/controllers/index.js
@@ -2,7 +2,7 @@ const path = require('path');
 const fs = require('fs');
 
 const basename = path.basename(module.filename);
-const routes = {};
+const controllers = {};
 
 fs
   .readdirSync(__dirname)
@@ -11,7 +11,42 @@ fs
     const fullPath = path.join(__dirname, file);
     const fileName = file.substring(0, file.length - 3);
 
-    routes[fileName] = fullPath;
+    controllers[fileName] = fullPath;
   });
 
-module.exports = routes;
+
+/**
+ * @class Controller
+ * This class will be an extendable factory for other controllers
+ **/
+class Controller {
+  constructor () {
+    // TODO:
+  }
+
+  /**
+   * @method messageResponse
+   * @param {String} msg - a message for the given response
+   * @param {Object} data - will attach any object as data,
+   *                        will default to an empty array.
+   * @return {Object}
+   */
+  messageResponse(msg, data = []) {
+    const response = {
+      'message': msg,
+      'data': data,
+    };
+
+    return response;
+  }
+
+  /**
+   * @param {Request} req - Express request object
+   * @param {Response} res - Express response object
+   */
+  methodNotAllowed(req, res) {
+    res.status(405).json(this.messageResponse('Method Not Allowed'));
+  }
+}
+
+module.exports = {Controller, controllers};
